Use recent version without comment in query
The 'someone built this now' comment is no longer ignored in choosing
the redirection answer.  No redirection is returned when that extra text
is included in the query.

Update from 2.46 to the most recent Jenkins long term support release,
2.204.x, so that the returned version matches with the passed in argument.
That will allow the test to run successfully for roughly 9 months before
the next change.