Use checkout instead of git step
On my test environment, the git step generates a warning that the checkout
will not be done inside the docker container. It also fails to find the
change.  The git step has a number of tests that confirm it continues to
work in typical scenarios.  It should be considered a deprecated method.
The checkout method is the preferred replacement.