Add a comment on the java check
It checks the path of the Java runtime reported by `mvn -v`.  If
the path includes the expected string, the test passes.  If the path
does not include the expected string, the test fails.

I had been expecting a test that checks the java version number as
reported by `mvn -v`.